pyproject.toml
[project]
name = "fastapi-web-reader"
version = "0.1.0"
description = "A simple API to read books and your own library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "alembic==1.16.1",
    "alembic-postgresql-enum==1.7.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "click==8.2.1",
    "cryptography==45.0.4",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi-cli[standard]==0.0.7",
    "fastapi[standard]==0.115.12",
    "frozenlist==1.7.0",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.5.0",
    "nicegui>=2.20.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "propcache==0.3.2",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.6",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pyjwt[crypto]==2.10.1",
    "pytest==8.4.0",
    "pytest-asyncio==1.0.0",
    "python-dotenv==1.1.0",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytoolconfig[global]==1.3.1",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "rope==1.13.0",
    "rsa==4.9.1",
    "ruff==0.11.13",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uvicorn[standard]==0.34.3",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "yarl==1.20.1",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D100", "D101", "D104", "D107", "D203", "D213"]
fixable = ["ALL"]

